refactor: split organization CLI into domain-specific files
Move org team, robot, application, marketplace, quota, auto-prune, and proxy-cache commands out of the organization.go monolith so cmd/ mirrors the lib/ layout.
